fix countdown
remove timer from off-thread
put timer in main thread
fix tests
remove worker dependencies
fix docs
