+++ jest.config.js
@@ -16,7 +16,6 @@ module.exports = {
   ],
   moduleNameMapper: {
     "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/test/__mock__/fileMock.js",
-    "^(worker-plugin\/loader)+.*": "<rootDir>/test/__mock__/fileMock.js",
     "\\.(css|less|scss|sass)$": "<rootDir>/test/__mock__/styleMock.js",
   },
   rootDir: process.cwd(),
