+++ package.json
@@ -104,8 +104,7 @@
     "rollup-plugin-typescript2": "^0.19.2",
     "rollup-plugin-url": "^2.1.0",
     "semver": "^7.1.3",
-    "shelljs": "^0.8.3",
-    "worker-plugin": "^4.0.3"
+    "shelljs": "^0.8.3"
   },
   "dependencies": {
     "@changesets/cli": "^2.9.0",
