+++ packages/ui/packages/Countdown/index.mdx
@@ -9,14 +9,8 @@ import Countdown from './src/index';
 
 # Countdown
 the countdown component is a useful countdown that gives you customization ability to implement the count down however you want.
-the core of the countdown has been implemented in the worker and it communicates with the component using the psotmessage.
-also the remaining time renders using the render method of the react-dom instead of updating the state, so there's no worry about the re-rendering  overheads.
+the remaining time renders using the render method of the react-dom instead of updating the state, so there's no worry about the re-rendering overheads.
 
-<br />
-
-# Import
-because of the worker, unlike the other packages, the countdown should be imported this way.
-<code>import Countdown from '@snappmarket/ui/Countdown'</code>
 
 <br />
