+++ packages/ui/packages/Countdown/src/index.js
@@ -2,9 +2,6 @@ import * as React from 'react';
 import {useEffect, useRef} from "react";
 import PropTypes from 'prop-types';
 
-// eslint-disable-next-line import/no-unresolved
-import countdownWorker from 'worker-plugin/loader?name=countdownWorker&esModule!./worker';
-
 import {render} from "react-dom";
 import { StyledCountdownWrapper, StyledCountDownItem } from './styles';
 
@@ -15,54 +12,60 @@ const Countdown = ({ date, className, children, onStart, onEnd }) => {
    * initiate the countdown with empty value till worker registers
    */
   useEffect(() => {
-    handleRenderCountdown({
-      days: "",
-      hours: "",
-      minutes: "",
-      seconds: "",
-    });
+    const remaining = date - Date.now()
+    handleRenderCountdown(remaining > 0 ? remaining :0);
   }, []);
   /**
    * registers the worker when the wrapper has been rendered
    */
   useEffect(() => {
     if(countdownRef.current) {
-      handleRegisterWorker();
+      onStart(Date.now());
+      handleStartCountDown();
     }
-  }, [countdownRef.current])
+  }, [countdownRef.current]);
 
-  /**
-   * creates an instance of worker from countdownWorker that has been parsed by worker-plugin
-   */
-  // @todo: use comlink instead to be able to expose the count down calculation, so the remaining time will be visible before the initialization
-  const handleRegisterWorker = () => {
-    // eslint-disable-next-line no-undef
-    const worker = new Worker(countdownWorker);
-    worker.postMessage({ key: 'setDate', value: date });
-    worker.onmessage = ({data}) => handlePostMessage(data);
-  }
-
-  /**
-   *
-   */
-  const handlePostMessage = ({key, value}) => {
-    const lookup = {
-      countdown: handleRenderCountdown,
-      start: onStart,
-      end: onEnd,
-    }
-    if(lookup[key])
-      lookup[key](value)
+  const handleStartCountDown = () => {
+    const interval = setInterval(() => {
+      const now = Date.now();
+      const remaining = date - now;
+      if(remaining <= 0) {
+        clearInterval(interval);
+        handleRenderCountdown(0);
+        onEnd(now);
+      }
+      else {
+        handleRenderCountdown(remaining);
+      }
+    }, 1000);
   }
 
   /**
    * renders the countdown with given or default template without re-rendering the entire component
    */
-  // @todo: add ability to render styled component in children
-  const handleRenderCountdown = countdown => {
+  const handleRenderCountdown = remaining => {
+    const countdown = handleCalculateCountDown(remaining)
     if(countdownRef.current) {
       const template = children(countdown) || handleDefaultTemplate(countdown);
       render(template, countdownRef.current);
+    }
+  }
+
+  /**
+   * calculates the remaining time to the due date
+   * @returns {{hours: number, seconds: number, minutes: number, days: number}}
+   */
+  const handleCalculateCountDown = remaining => {
+    const second = 1000;
+    const minute = (second * 60);
+    const hour = (minute * 60);
+    const day = (hour * 24);
+
+    return {
+      days: Math.floor(remaining / day),
+      hours: Math.floor((remaining % day) / hour),
+      minutes: Math.floor((remaining % hour) / minute),
+      seconds: Math.floor((remaining % minute) / second),
     }
   }
